+++ packages/lumx-react/src/components/post-block/Demos.stories.tsx
@@ -1,0 +1,3 @@
+export default { title: 'LumX components/post-block/Demos' };
+
+export { App as Default } from './demos/default';